Remove the systemd source to avoid confusion #13
The source code was copied to the repository to study the text output
on the command line of systemd. These C files were not needed for the
functionality of the plugin. Furthermore, it confused the programming
language detection on Github: Shell 52.0% C 33.7% Python 14.3%
The deleted files are replaced with links to the systemd’s repository
on Github.