Get rid of README-template.md
The current help text of the command line interface is now patched
into the README.md file without the help of the README-template.md
file. Now there are on the one hand less files in the root directory,
on the other hand overwriting errors are avoided. Changes were not
applied when directly editing the README.md file and not the
README-template.md.